requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
openai==1.3.5
tiktoken==0.5.1
duckdb==0.9.2
pydantic==2.5.2
python-dotenv==1.0.0
textstat==0.7.3
semantic-text-splitter==0.22.0
tqdm==4.66.1
PyYAML==6.0.1
jinja2==3.1.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0
